changed the member of BoxedInt and LitInt from Integer to long.

+++ brt/src/jp/ne/sakura/uhideyuki/brt/brtsyn/BoxedIntObj.java
@@ -1,7 +1,7 @@
 package jp.ne.sakura.uhideyuki.brt.brtsyn;
 
 public class BoxedIntObj extends HeapObj {
-    public Integer value;
+    public long value;
     public BoxedIntObj(LitInt x){ value = x.value; }
     public String inspect(){
 	return String.valueOf(value);
